3d mode reworks/fixes
Changed the way wall/flat lighting on 3d floors is calculated to something a bit closer to how *zdoom does it, plus fixed a bunch of smaller issues with map geometry generation

+++ src/SLADEMap/MapObject/MapSector.cpp
@@ -624,7 +624,7 @@ uint8_t MapSector::lightAt(SectorPart where) const
 	}
 
 	// Clamp light level
-	return std::clamp<u8>(light, 0, 255);
+	return std::clamp<int>(light, 0, 255);
 }
 
 // -----------------------------------------------------------------------------
